[Issue #389] Add changeset to trigger version bump and deploy (#438)
* add changeset to trigger version bump and deploy

* fix: Changeset parsing issue due to package override
In a previous commit we overrided a vulnerable version of js-yaml.
This override wasn't supported by the version of changelog we were using
which caused `pnpm changelog` to raise a parsing error. We fixed this
by bumping to the latest patch version of changelog, which supports a
non-vulnerable js-yaml version which means we can remove the override

---------

Co-authored-by: widal001 <billy.daly@agile6.com>